feat: P1 trend stage analysis + support/resistance levels
Backend:
- Add models/trend.go: TrendStage, SupportResistance, TrendAnalysis, MultiTrendStock, MultiTrendResponse, CorrelationResponse, PeriodIndicators
- Add services/trend_analyzer.go: trend direction/stage detection, S/R level calculation
- TrendStage: direction(上升/下降/震荡), stage(早期/中期/末期), strength, confidence, signals
- SupportResistance: 3 support + 3 resistance levels from MAs and Bollinger bands
- Integrate trend analysis into analyzeSingleWithMode

Frontend:
- Add trend_analysis to RawAnalyzeResponse and AnalyzeResult interfaces
- Add TrendAnalysisCard component to AnalyzePage
- Display: direction, stage, strength, confidence, S/R levels, key MAs, nearest level distance

Trend detection logic:
- MA arrangement (30%), MACD signal (25%), MACD histogram trend (15%), RSI level (15%), volume confirmation (15%)
- Stage: early (just crossed MAs), mid (moderate distance), late (far from MAs + momentum decreasing)
- S/R levels from MA5/10/20/60 + Bollinger upper/mid/lower bands

+++ internal/handlers/trend.go
@@ -185,7 +185,7 @@ func (h *TrendHandler) analyzeMultiTrend(ctx context.Context, code string) *mode
 	weeklyIndicators := computePeriodStrength(weeklyKlines)
 	monthlyIndicators := computePeriodStrength(monthlyKlines)
 
-	overallStrength := (dailyIndicators.Strength + weeklyIndicators.Strength + monthlyIndicators.Strength) / 3
+	overallStrength := float64((dailyIndicators.Strength + weeklyIndicators.Strength + monthlyIndicators.Strength)) / 3.0
 
 	return &models.MultiTrendStock{
 		Code:            code,
